✨ Feat: `Notification` type 추가
